Creo que ya :)

+++ application/views/admin/permisos/list.php
@@ -15,7 +15,9 @@
             <div class="box-body">
                 <div class="row">
                     <div class="col-md-12">
+                      <?php if ($permisosValidar->insert == 1):?>
                         <a href="<?php echo base_url();?>administrador/permisos/add" class="btn btn-primary btn-flat"><span class="fa fa-plus"></span> Agregar Permiso</a>
+                      <?php endif; ?>
                     </div>
                 </div>
                 <hr>
@@ -72,8 +74,12 @@
 
                                             <td>
                                               <div class="btn-group">
-                                                <a href="<?php echo base_url()?>administrador/permisos/edit/<?php echo $permiso->id;?>" class="btn btn-warning"><span class="fa fa-pencil"></span></a>
-                                                <a href="<?php echo base_url();?>administrador/permisos/delete/<?php echo $permiso->id;?>" class="btn btn-danger"><span class="fa fa-remove"></span></a>
+                                                <?php if ($permisosValidar->update == 1):?>
+                                                  <a href="<?php echo base_url()?>administrador/permisos/edit/<?php echo $permiso->id;?>" class="btn btn-warning"><span class="fa fa-pencil"></span></a>
+                                                <?php endif; ?>
+                                                <?php if ($permisosValidar->delete == 1):?>
+                                                  <a href="<?php echo base_url();?>administrador/permisos/delete/<?php echo $permiso->id;?>" class="btn btn-danger"><span class="fa fa-remove"></span></a>
+                                                <?php endif; ?>
                                               </div>
                                             </td>
                                         </tr>
